Move analysis/_linear.py to _regrid.py
