+++ lib/iris/tests/results/analysis/regrid/both_circular.cml
