+++ lib/iris/tests/results/analysis/regrid/both_circular.data.0.npy
