+++ lib/iris/tests/results/analysis/regrid/circular_grid.data.0.npy
