+++ lib/iris/tests/results/analysis/regrid/circular_src.data.0.npy
