+++ lib/iris/tests/results/analysis/regrid/linear_masked_altitude.cml
