+++ lib/iris/tests/results/analysis/regrid/linear_masked_altitude.data.0.npy
