+++ lib/iris/tests/results/analysis/regrid/linear_partial_overlap.cml
