+++ lib/iris/tests/results/analysis/regrid/linear_subset.data.0.npy
