+++ lib/iris/tests/results/analysis/regrid/linear_subset_anon.data.0.npy
