+++ lib/iris/tests/results/analysis/regrid/linear_subset_masked_1.cml
