+++ lib/iris/tests/results/analysis/regrid/linear_subset_masked_2.data.0.npy
