+++ lib/iris/tests/results/analysis/regrid/nearest_masked_altitude.cml
