+++ lib/iris/tests/results/analysis/regrid/nearest_masked_altitude.data.0.npy
