+++ lib/iris/tests/results/analysis/regrid/nearest_partial_overlap.cml
