+++ lib/iris/tests/results/analysis/regrid/nearest_partial_overlap.data.0.npy
