+++ lib/iris/tests/results/analysis/regrid/nearest_subset.data.0.npy
