+++ lib/iris/tests/results/analysis/regrid/nearest_subset_anon.data.0.npy
