+++ lib/iris/tests/results/analysis/regrid/nearest_subset_masked_1.cml
