+++ lib/iris/tests/results/analysis/regrid/nearest_subset_masked_1.data.0.npy
