+++ lib/iris/tests/results/analysis/regrid/nearest_subset_masked_2.data.0.npy
