+++ lib/iris/tests/results/analysis/regrid/no_overlap.data.0.npy
