+++ lib/iris/tests/results/analysis/regrid/non_circular.cml
