+++ lib/iris/tests/results/analysis/regrid/non_circular.data.0.npy
