+++ lib/iris/tests/results/visual_tests/test_regrid.Test___call____visual.test_osgb_to_latlon.0.png
