+++ lib/iris/tests/results/visual_tests/test_regrid.Test___call____visual.test_subsample.0.png
