+++ lib/iris/tests/unit/analysis/regrid/__init__.py
@@ -1,4 +1,4 @@
-# (C) British Crown Copyright 2014, Met Office
+# (C) British Crown Copyright 2014 - 2015, Met Office
 #
 # This file is part of Iris.
 #
@@ -14,6 +14,6 @@
 #
 # You should have received a copy of the GNU Lesser General Public License
 # along with Iris.  If not, see <http://www.gnu.org/licenses/>.
-"""Unit tests for the :mod:`iris.analysis._linear` module."""
+"""Unit tests for the :mod:`iris.analysis._regrid` module."""
 
 from __future__ import (absolute_import, division, print_function)
